+++ mrv2/lib/mrvApp/mrvApp.cpp
@@ -1210,8 +1210,10 @@ namespace mrv
 #if defined(TLRENDER_FFMPEG)
         out["FFmpeg/YUVToRGBConversion"] = string::Format("{0}").arg(
             p.settings->getValue<int>("Performance/FFmpegYUVToRGBConversion"));
-        out["FFmpeg/FastYUV420Conversion"] = string::Format("{0}").arg(
-            p.settings->getValue<int>("Performance/FFmpegColorAccuracy"));
+        int fastYUV420PConversion =
+            !(p.settings->getValue<int>("Performance/FFmpegColorAccuracy"));
+        out["FFmpeg/FastYUV420PConversion"] =
+            string::Format("{0}").arg(fastYUV420PConversion);
         out["FFmpeg/ThreadCount"] = string::Format("{0}").arg(
             p.settings->getValue<int>("Performance/FFmpegThreadCount"));